requirements.txt
streamlit>=1.28.0
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.4
Pillow>=10.0.0
mediapipe==0.10.32
absl-py>=1.4.0
attrs>=25.4.0
flatbuffers>=25.12.19
matplotlib>=3.10.0
protobuf>=4.25.3
opencv-python-headless==4.9.0.80
